updated seed script and separated concerns
