+++ docker-compose.yml
@@ -1,0 +1,11 @@
+version: '3'
+
+services:
+
+  db:
+    image: mysql:5.7
+    environment:
+      MYSQL_ROOT_PASSWORD: 
+
+  database:
+    image: mysql:5.7
