+++ package.json
@@ -13,8 +13,8 @@
     "test": "jest",
     "coverage": "jest --coverage",
     "build": "webpack -d --watch",
-    "server": "nodemon server/index.js",
-    "seed": "node server/index.js",
+    "start": "nodemon server/index.js",
+    "seed": "node db/seedScript.js",
     "schema": "mysql -uroot < db/schema.sql"
   },
   "repository": {
