+++ scratch.txt
@@ -32,3 +32,20 @@
     text-decoration: none;
     font-size: 8px;
   `;
+
+
+   - - - 
+
+     monthlycost:
+    build: .
+    depends_on:
+      - 'database'
+    ports: 
+      - '8080:8080'
+
+  # FILLMEIN:
+  #   build: results_server/
+  #   depends_on:
+  #     - 'database'
+  #   ports:
+  #     - '3000:3000'
